Bug #916: Change RTSP TCP/UDP negotiation to fix ACTi
The previous method of issuing a SETUP for TCP, and then a SETUP
for UDP if that fails would break ACTi devices, in a way that is
difficult for us to recover from (it would require some curl reset).

Instead, we do what the RTSP specification intends: Send one SETUP,
with the Transport header containing a prioritized list of transports,
and allow the server to choose. This avoids the ACTi issue and does
not harm support for devices that only allow one of UDP or TCP.

Conforming servers should follow the client's priorities, but naturally,
at least one (Grandstream) does not. After the patch, Grandstream
devices - which do support TCP streams - will set up UDP regardless of
our priority. I don't believe this is a functional regression significant
enough to prevent this change: We support UDP anyway, and this method
is by the specification, which will hopefully avoid future issues.