Bug #914: Attempt to slightly reduce memory leakage in Apache
This is an unproven partial workaround that might help lower memory
usage during long mjpeg.php requests, by downgrading the connection
to HTTP 1.0 and thus avoiding the use of the chunked transfer encoding
(which requires significant server-side processing). This does not
harm functionality, but I haven't confirmed that it reduces the memory
leak either.