Add rtsp-record to utils
This is a simple, ugly tool to record a RTSP stream to a file
using the ffmpeg RTSP implementation, intended as a test to see
if we can use this for the server.

May require ffmpeg 0.7 for servers that require authentication.
A workaround is included to avoid depending on GPL-licensed libx264.

Confirmed to work so far with Axis and Grandstream devices, others
pending.