Persist RTP connections between media resets
This removes a potentially large delay between continuous recordings
from RTP, as well as avoiding a few related issues.

We keep a PTS offset for RTP packets to ensure that each recording
starts from a PTS of 0. Recordings must start with a video keyframe,
so we drop all packets after the end of a recording until the next
of those. In the future, we should find a way to sync audio and video
at the recording cut, possibly by offsetting the actual end until
just before a keyframe and performing audio magic.