Improve behavior of audio PTS on RTP streams
The adjustment done during PTS resets (i.e. at the beginning of a
new recording) must be scaled to each individual stream, since they
are not all kept in the same time_base.

We now don't update the PTS interval of a stream to less than 1/4th
of its current value, unless two consecutive frames are below that
threshold. This helps prevent false positive forward jumps on audio
streams, when our interval drops dramatically for one frame, based
on the observation that this should not occur twice consecutively.
Recovery from the bad forward jump situation will be necessary as
well.

Drop negative PTS frames in bc_av_out. These can have some legitimacy,
if part of an audio packet contains the time of the first frame in
a recording, but are more likely to confuse some players. This can
result in a small audio gap in the beginning of the recording, but will
not cause desynchronization. We'll have a better solution to this when
recording splits are buffered to avoid the keyframe gap.

With this commit, Axis and ACTi test devices seem to record proper
audio under good network conditions.