Redesign device configuration updates for thread safety
The old design would modify shared variables and even more complex
state (such as driver calls) from the main thread while that data
is in active use by the device thread. This was a huge opportunity
for race conditions that could confuse or corrupt recordings, or
cause crashes.

It is desirable to avoid running SQL queries on device threads.
So, we poll the database from the main thread as before, and
compare its new configuration to a copy of the current device
configuration. If changes are made, the new configuration is copied
to a temporary area and a flag is set. The device thread checks
this flag, and applies configuration changes as a result. This
allows thread-safe application of the changes, without putting a
mutex in the common case of the device thread.

This may fix the solo video corruption trigger from frequent
recording resets, and possibly the uptime-related video corruption
trigger as well.