Refactor bc_open_avcodec to not depend on the current frame
Previously, we would decode the current frame to get some stream
parameters (size and framerate) and to save the snapshot JPEG.
Snapshots are currently disabled, and we can get those parameters
without decoding for V4L2 and RTP devices.

This is necessary for motion prerecording, where the current
frame has no relation to the start of the new recording.

Reverts 5111d9c5 and 600774a2