Bug #1009: Create the logfile during postinst
When syslog is running as the 'syslog' user, it might not be able
to create new files under /var/log, and definitely not as with
'bluecherry' group. Logrotate can handle these fine, so all we
need to do is to create it as root at install-time. This should
fix issues on Ubuntu server installations.