Bug #1021: Add room for null termination on the per-device schedule
Previously, the variable was exactly as large as its data; strlcpy
then did the right thing and truncated to set the last field to 0.
This resulted in the schedule defaulting to continuous at the last
hour of the schedule, which is 11pm saturday.

Fixed by null terminating both schedules (as I feel it's cleaner),
and making sure they have enough space.