Bug #1026: Strip B2 headers for ACTi streams
As noted in more detail on the bug, these headers are not conformant
with the MPEG4 specification, and confuse parsers, occasionally
resulting in a false VOL header that breaks the stream. Since it isn't
possible for a parser to handle these generically, we can strip them
in the server before recording or decoding packets. This should fix
the corruption issues on ACTi devices entirely.