Bug #1041: Use driver information to determine solo video codec
We cannot rely on the database to always be perfectly accurate about
whether a device is solo6010 or solo6110; in at least one instance,
that has been wrong and resulted in bad recordings.

We can get this information (badly, but still) from the driver for
the device when opened instead, which is guaranteed to be correct.