Bug #912: Drop privileges in bc-server instead of a helper script
Upstart is particularly bad at tracking processes that require a
helper script to run, e.g. our use of sudo (or su, or
start-stop-daemon) to drop privileges. This resulted in the upstart
start/stop commands hanging on some systems.

By dropping privileges directly in the process instead of with a
helper, we can work around upstart's incompetence here.