Refactor camera events
This code was (and somewhat still is) a badly designed disaster.
The pointless abstraction between media and events, numerous bugs,
slow and incorrect SQL behavior, etc, made it a huge source of
problems in a component that needs to be very robust.

This is a dramatic refactor to simplify the whole process of
creating events/media related to cameras. It fixes a handful of
related bugs (no tickets I know of), and allows for specifying a
start time in the past, which is needed by #1025.