Feature #1049: Use motion prerecord as a post-event buffer
Instead of ending a motion event [postrecord] seconds after the last
frame with motion, stop writing frames, but keep the event and the
recording open until a packet that was not included in this recording
is dropped from the prerecording buffer. If a new frame with motion is
received before we've lost a frame, dump frames from the prerecord
buffer to catch up and continue the existing event and recording from
there.