Add bc_status_component_{begin,end,error} to track critical status
Components defined in an enum in bc-server.h can report status by
wrapping blocks of code in _begin and _end, and calling _error with
a message for any serious problems that occur. This is intended to
be used only for errors that will affect normal operation of the
entire server. The database is updated every minute with a timestamp
reflecting the last known time that the server was operational, and
an error message if applicable. As per #1004, it's considered to be
broken if the timestamp is older than some interval *or* there is
an error message set.

This isn't a perfectly designed nor perfectly implemented solution,
but it should suffice for the requirements of feature #1004 for now.