Detect solo devices by PCI ID instead of driver
This allows us to enumerate all solo devices on the system, including
those not yet initialized by the solo6x10 driver (which can sometimes
be slow on systems with many cards). As a result, we can wait until
after all cards are initialized before doing device detection, and
avoid the confusing situation where some cards disappear for a few
minutes after boot.

Changes to how detection is run to handle that delay will be coming
shortly.

This also replaces the usage of libsysfs with libudev, which is
cleaner and preferable (libsysfs is deprecated). A dependency for
libudev is added, the dependency for libsysfs is removed, as well
as the associated -dev packages.

Bug #1066