Retry solo detection when devices aren't initialized yet
We'll warn if there are devices not initialized at startup, and
trigger an error if they're still not ready after 15 seconds.
No recording threads will be started for solo devices until
detection has completed successfully. After the first successful
detection, we go to an every-two-minutes schedule to check for
changes (as it did before); though honestly, that should only
matter if someone is hotplugging.

Bug #1066