Properly destroy rtsp_stream instances and end their sessions
We rely on the fact that only one arbitrary thread (the recording thread)
and the RTSP server thread will interact with a rtsp_stream; to end a
stream, the recording thread flags it, wakes the RTSP thread, and ceases
any use. The RTSP thread, when woken, will collect these dead streams and
properly free them as well as ending their sessions.

Because there's no way to signal the end of a particular session in RTSP,
other than RTCP BYE packets (which are a good idea, but not worth dealing
with right now), we just close the associated RTSP connection when a session
is ended server-side. No current use case involves more than one stream per
RTSP connection, so this shouldn't be of concern.

This also relies on the assumption that there is only one RTSP thread, but
that assumption should hold true even if we end up using multiple listeners.