Authentication for RTSP streams
This is just Basic HTTP authentication on the DESCRIBE and SETUP
commands; commands operating on an existing session don't authenticate
to avoid the locking and queries. It would serve little purpose, anyway.

Unfortunately, we can't use Digest authentication, because passwords are
stored as salted MD5. We'd need to store a separate hash for Digest, or
a plaintext password. RTSP-over-SSL might be another idea for future
security, but the performance impact for interleaved streams could be
significant.

A third option for increased security would be an access token system,
where the client requests one-time-use auth tokens from the webserver
and sends those to set up RTSP streams. This would be annoying to
implement, but much more secure than what's being done now (which amounts
to the username and password being sent in plaintext).