Allow keyframe-only streaming using a RTSP URL parameter
Due to clients' incorrect behavior with URL query options, this
option is passed like streamid, in /-delimited key=value pairs as
part of the URL. The path for a keyframe only stream is
/live/x/mode=keyframe, where x is the device ID.

This should be made modifiable on an existing stream using
SET_PARAMETER at some point in the future, but doing so clientside
is non-trivial. There are problems with some streams and the current
implementation for dropping frames, but a fix for that is forthcoming.