Don't ignore media without a recorded end or size for cleanup
As per #1091, among other situations, some media would never be
updated with an end timestamp or a filesize, especially when it
was in progress during a server shutdown. These files would never
be seen by cleanup, and slowly eat away at available space.

There are other fixes to this bug that we need as well, but this
is the most important: Any media we record should be elgible for
cleanup, unless the user has decided to archive it.