Improve buffering on RTSP connections for TCP streaming
As per Bug #1097, we can avoid corrupting the client's image by
dropping packets from the buffer more intelligently. Truncated
packets can break the connection or cause error concealment to
ruin the image for the client.

This ugly code ensures that buffers and frames stay intact as
sent to the client, and trims itself of unnecessary frames when
a new reference frame is available. If the buffer exceeds 2MB
without any further trimming possible, the connection is closed.