Fix improper trim & memory leak in rtsp_write_buffer
Trimming on partial reference frames would result in the beginning
of a new reference being discarded, which ruined the stream.

Additionally, a mistake in the trimming logic would leak buffers
from the first frame during a trim, and corrupted the stream by
sending a truncated frame immediately before a trimmed reference.

Bug #1103, fix for regressions in bug #1097