Don't attempt to connect to MySQL without user/pass/db
Due to the call to get_handle when opening the database, there would
be an attempt to connect with the default settings (the system username,
no password, no database); if that succeeded, the proper settings would
never be used, and all queries would fail. A MySQL configuration change
in Ubuntu 12.04 results in these connections succeeding.

Bug #1112