Update duration on in-progress non-media events at startup
This is done to prevent 'not found' events and other non-media
event types from polluting results by appearing to be indefinitely
active. The duration is set based on the last known running time
of the server, as determined by the ServerStatus table, which should
be accurate within a few minutes.