Remove old ALSA/solo audio code
This hasn't been functional for a long time, and it never worked
properly due to trying to juggle ALSA and solo calls in the same
thread. It also conflicted badly with pulseaudio and was otherwise
generally broken.

Removed all of the remenants of this implementation. Solo audio can
be implemented later, properly, as part of the library's v4l2device.