Bug #1161: Increase the libav log threshold for running streams
As per the bug, sometimes libav errors can be emitted in the process of
receiving frames or decoding despite not causing any trouble, and this
confuses customers.

After a RTP device is started, raise the threshold to FATAL, unless debug
mode is enabled. This may hide some legitimate errors about connection
loss or missed packets; we'll have to evaluate that in practice and see
if we need to do anything more.