Force single-threaded decoding for RTP
When using motion detection, multi-threaded decoding would automatically
be used for some codecs. Since this system is often handling several
concurrent streams, and needs to avoid stealing too much CPU time for
any one device, single-threaded decoding should always be used.

This could result in an increase in lag/corruption for IP motion
streams that would exceed the power of a single CPU core. That case is
already a serious problem, so I'm not too concerned about it for now.
The buffered architecture will eventually allow motion to run at a
different rate and without blocking any other part of the system.

This fixes a problem with decoding errors not being displayed with the
associated device name or subjected to the right avlib log threshold.