Implement real stream_properties in lib
stream_properties now holds everything necessary to decode or record
the stream, without directly interacting with the input. It holds
similar information as AVCodecContext.

Because this information may change during the stream, and it's important
that those changes take effect at the right time, each stream_packet holds
a std::shared_ptr to a stream_properties instance for its current properties.
The class is immutable after being passed in this manner; for any changes, a
new instance should be created and included with all packets from when it began
to apply. Users can test for changes by comparing these pointers, or the data
itself.

There is some further room for combining code, especially related to decoding
and JPEG snapshots, as a result of this.