Carefully avoid deadlocking stream_elements on destruction
This is not the most robust solution, but given that long term plans
are to move this system away from lock+wait+one-thread-per-element, it's
all that seems worth implementing now. The disconnect/destroy sequence is
carefully managed now to ensure that consumers are disconnected before
the connected source is destroyed, and none will deadlock.

Signed-off-by: Ismael Luceno <ismael.luceno@corp.bluecherry.net>