Implement smarter logging, including log levels and thread-based contexts
Logging now has a level, defaulting to Info, which can be controlled with
a commandline parameter or on a per-device basis using the debug_level field
in the database (which retains it's old on/off usage).

log_context represents a context for log messages, such as a single recording
device, which may (but is not required to) have its own minimum level, and
prefixes messages with a context name. It uses reference-counted data to allow
safe usage across threads.

You can log directly to a context with its log() function, or use the bc_log
function, which uses a thread-specific context from a stack controlled by
bc_log_context_push and bc_log_context_pop. A context can be pushed to that
stack to put subsequent messages from that thread into that context, and popped
after that operation has completed.

Signed-off-by: Ismael Luceno <ismael.luceno@corp.bluecherry.net>