BUG 1207: Destroy connection when live connection is broken
For the bug in question 1207, after several hours the customer
faces an outage in the service. Per research it could lead to
a bad handling of an error from the network in the server side,
so from now (this patch) if it face the same problem, it will
close the stream and clean it up properly.

This problem is not faced by other customers, its hard to reproduce.

Signed-off-by: Eduardo Silva <eduardo@corp.bluecherry.net>