www: edit ip cam: hide manufacturer & model fields
This change is done to avoid confusion.
Currently we don't store model information in DB, so all cameras look like they're configured for the first manufacturer in DB ("3S Vision").
This is more a hack than an ideal solution to the situation. My excuse is lack of task force in the project team.