echo 2050300 > misc/sql/installed_db_version
This value represents 2.5.3 version number which is relevant to expected
time of this feature being merged to master.
General policy of value of this file is "first stable release in which
the DB schema upgrade happens". But it is not expected to have this
feature in 2.4 branch, due to unstable nature of it. We don't mind to
make 2.6 feature if it is considered stable.