sql: make foreign keys to Devices/id as cascade
This commit is changing the device_id references as cascade.

This is allowing to reorder cameras internally using simple
update, like e.g.:
update Devices set id=5 where id=4;

As the bluecherry server is not allowing (yet? ;)) to change
this id (and thus related directory name/id in nvr structure),
it is the easy way to reorder cameras without the need to
delete it (and its settings) and re-add again.