logrotate: drop "create"
In Debian, even with rsyslog installed, there's no "syslog" user, and
rsyslog runs as root.
It is expected that rsyslog will create the file itself.
There are no particular automated use of log files, so it doesn't matter
for files to have "bluecherry" group and be readable for that group.