Fix video snapshots saving
media_writer::scale_frame() converts the frame to a JPEG-compatible
pixel format, but it doesn't carry over all the frame metadata like
format, and picture dimensions. This is the apparent reason why video
snapshots don't work. This problem has been reported by many users after
the 3.1.2 release, which brought the upgrade of bundled ffmpeg to
6.1.1 from 4.2.1.

This change has been tested.

The problem shows up in the log like this:

    I(1/camera_name): Making a snapshot
    E(1/camera_name): avcodec_send_frame: snapshot encoding failed: Invalid argument
    E(1/camera_name): failed to write snapshot frame
    E(1/camera_name): Feeding snapshot decoder failed
    E(1/camera_name): Failed to make snapshot