Improve timestamps handling at recording muxing
Specifically:

- Restart recording on timestamping problems
- Reset last_mux_dts[] on media_writer::close()
- consistently use AV_NOPTS_VALUE as "NULL" for last_mux_dts
- don't update last_mux_dts to a bad one which isn't muxed

Reliability regressions were reported on certain setups attributed to
changes in 3.1.2 and 3.1.6.

One scenario in which problems arise is: an HDMI-to-RTSP encoder is
connected as a device to Bluecherry server, and the upstream video
source connected to the encoder is switched off and later back on.

This commit solves the issues for the scenario which we reproduced in
our lab.