Revert "Shut down the server orderly on termination signals"
This reverts commit 49679320eec91114635549d9a34774db757413ac.

Shutting down is important and we want to get it right as soon as
possible, but such change is evidently destabilising to the service
behaviour, requiring many follow-up fixes and careful testing and
observation. We are taking our time and will ship this in a future
release separately from the changes introduced since 3.1.7.