Drop postinst code "deoccupying root crontab"
The complaints were received that the code keeps removing entries
deliberately put into root crontab which match the regex.
(Must admit the regex is indeed very broad.)

It was suggested to make matching more precise.
It is also an option to reframe our database schema upgrading mechanism
as a more general upgrade mechanism, and move that code there as it's
meant to be ran only one time.

But I believe just taking out the whole "cleanup" code is acceptable
because
- large amount of people have already upgraded and have run this code
  (with, hopefully, net-beneficial effect), and
- having any code manipulating on root crontab is irksome to a degree,
- all code incurs maintenance costs, and this is not a very useful piece
  of code going forward: we'll keep looking at it with eyeballs, and to
  keep the code and stay confident in it, we want to have that code
  tested. No code, no maintenance and testing costs and delays.